new : добавлен dockercompoyse

+++ main-service/src/main/java/ru/practicum/categories/service/CategoriesServiceImpl.java
@@ -2,19 +2,24 @@ package ru.practicum.categories.service;
 
 import lombok.RequiredArgsConstructor;
 import org.mapstruct.factory.Mappers;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
-import ru.practicum.categories.dto.CategoryDto;
-import ru.practicum.categories.dto.ResponseCategoryDto;
-import ru.practicum.categories.dto.NewCategoryDto;
+import ru.practicum.categories.CategoryDto;
+import ru.practicum.categories.NewCategoryDto;
+import ru.practicum.categories.ResponseCategoryDto;
 import ru.practicum.categories.mapper.CategoriesMapper;
 import ru.practicum.categories.model.Category;
 import ru.practicum.categories.repository.CategoriesRepository;
-import ru.practicum.user.exception.NotFoundException;
+import ru.practicum.events.model.Event;
+import ru.practicum.events.repository.EventsRepository;
+import ru.practicum.exception.NotFoundException;
+import ru.practicum.exception.RequestAlreadyExists;
 
-import java.util.Collections;
+import javax.transaction.Transactional;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -23,30 +28,45 @@ public class CategoriesServiceImpl implements CategoriesService {
     private final CategoriesMapper categoriesMapper = Mappers.getMapper(CategoriesMapper.class);
     private final CategoriesRepository categoriesRepository;
 
+    private final EventsRepository eventsRepository;
+
+    @Transactional
     @Override
     public ResponseCategoryDto addCategory(NewCategoryDto category) {
-        Category savedCategory =  categoriesRepository.save(categoriesMapper.toCategory(category));
-        System.out.println("savedCat= " + categoriesMapper.toDto(savedCategory));
+        Optional<List<Category>> catWithSameName = categoriesRepository.findByName(category.getName());
+        if (catWithSameName.get().size() != 0) {
+            throw new RequestAlreadyExists("Категория с таким именем уже существует: " + category.getName());
+        }
+        Category savedCategory = categoriesRepository.save(categoriesMapper.toCategory(category));
         return categoriesMapper.toDto(savedCategory);
     }
 
 
+    @Transactional
     @Override
     public void deleteCategory(Long catId) {
+        Optional<List<Event>> eventsWithSameCat = eventsRepository.findAllByCategoryId(catId);
+        if (eventsWithSameCat.get().size() != 0) {
+            throw new RequestAlreadyExists("Нельзя удалить категорию, если с ней связаны события " + catId);
+        }
         Category category = categoriesRepository.findById(catId).orElseThrow(() ->
                 new NotFoundException("Такой категории нет " + catId));
-        System.out.println("category to delete= " + category + category.getId());
         categoriesRepository.deleteById(catId);
     }
 
 
+    @Transactional
     @Override
     public ResponseCategoryDto update(CategoryDto category, Long catId) {
         Category savedCategory = categoriesRepository.findById(catId).orElseThrow(() ->
                 new NotFoundException("Такой категории нет " + catId));
-        savedCategory.setName(category.getName());
-        Category updatedCategory = categoriesRepository.save(savedCategory);
-        return categoriesMapper.toDto(updatedCategory);
+        try {
+            savedCategory.setName(category.getName());
+            Category updatedCategory = categoriesRepository.save(savedCategory);
+            return categoriesMapper.toDto(updatedCategory);
+        } catch (DataIntegrityViolationException exception) {
+            throw new DataIntegrityViolationException("Такое имя уже есть: " + category.getName());
+        }
     }
 
 
@@ -57,9 +77,6 @@ public class CategoriesServiceImpl implements CategoriesService {
             final PageRequest pageRequest = PageRequest.of(page, size);
             Page<Category> result = categoriesRepository.findAll(pageRequest);
             List<Category> categories = result.getContent();
-//            if(categories.size() == 0) {
-//                return result.
-//            }
             return categoriesMapper.toListDto(categories);
         }
         List<Category> categories = categoriesRepository.findAll();
@@ -73,5 +90,4 @@ public class CategoriesServiceImpl implements CategoriesService {
                 new NotFoundException("Такой категории нет " + catId));
         return categoriesMapper.toDto(category);
     }
-
 }
